+++ grid_optimizer/grid.h
@@ -7,7 +7,7 @@
 #include <iostream>
 
 
-typedef std::vector<size_t> index;
+typedef std::vector<int> index;
 
 
 struct node {
@@ -25,17 +25,17 @@ struct node {
 class Grid {
 public:
 	Grid();
-	void read_file(const size_t IMAX, const size_t JMAX, const size_t KMAX, const std::string filename);
+	void read_file(const int IMAX, const int JMAX, const int KMAX, const std::string filename);
 	std::pair<alglib::real_2d_array, alglib::integer_1d_array> get_X_tags();
 	index get_ijk();
-	index get_ijk(const size_t & n);
-	node get_node(const size_t & i, const size_t & j, const size_t & k);
-	void set_node(const size_t & i, const size_t & j, const size_t & k, node p);
-	alglib::real_1d_array get_xyz(const size_t & i, const size_t & j, const size_t & k);
+	index get_ijk(const int & n);
+	node get_node(const int & i, const int & j, const int & k);
+	void set_node(const int & i, const int & j, const int & k, node p);
+	alglib::real_1d_array get_xyz(const int & i, const int & j, const int & k);
 private:
-	size_t IMAX;
-	size_t JMAX;
-	size_t KMAX;
+	int IMAX;
+	int JMAX;
+	int KMAX;
 	std::vector<std::vector<std::vector<node>>> values;
 };
